pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "md-drop"
version = "0.1.0"
description = "Capture Markdown content from anywhere into your Obsidian vault"
requires-python = ">=3.12"
dependencies = [
    "click>=8.0",
    "google-api-python-client>=2.0",
    "google-auth-oauthlib>=1.0",
    "markdownify>=0.14",
    "python-frontmatter>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
md-drop = "md_drop.sync:main"

[tool.hatch.build.targets.wheel]
packages = ["src/md_drop"]

[tool.hatch.build.targets.sdist]
include = ["src/md_drop"]
